requirements.txt
asgiref==3.8.1
boto3==1.35.25
botocore==1.35.25
dj-database-url==2.2.0
Django==5.1.1
django-extensions==3.2.3
django-heroku==0.3.1
django-role-permissions==3.2.0
django-stdimage==6.0.2
django-storages==1.14.4
jmespath==1.0.1
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
